requirements.txt
Flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.3
icalendar==6.1.2
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0